+++ src/main/java/me/marin/lockout/lockout/goals/advancement/SkewerMobsGoal.java
@@ -1,0 +1,33 @@
+package me.marin.lockout.lockout.goals.advancement;
+
+import me.marin.lockout.lockout.interfaces.AdvancementGoal;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.Items;
+import net.minecraft.util.Identifier;
+
+import java.util.List;
+
+public class SkewerMobsGoal extends AdvancementGoal {
+
+    private static final ItemStack ITEM_STACK = Items.NETHERITE_SPEAR.getDefaultStack();
+    private static final List<Identifier> ADVANCEMENTS = List.of(Identifier.of("minecraft", "adventure/spear_many_mobs"));
+
+    public SkewerMobsGoal(String id, String data) {
+        super(id, data);
+    }
+
+    @Override
+    public String getGoalName() {
+        return "Skewer 5 mobs at once";
+    }
+
+    @Override
+    public ItemStack getTextureItemStack() {
+        return ITEM_STACK;
+    }
+
+    @Override
+    public List<Identifier> getAdvancements() {
+        return ADVANCEMENTS;
+    }
+}