+++ src/main/java/me/marin/lockout/lockout/goals/kill/KillSlimeGoal.java
@@ -3,6 +3,8 @@ package me.marin.lockout.lockout.goals.kill;
 import me.marin.lockout.Constants;
 import me.marin.lockout.lockout.interfaces.KillMobGoal;
 import me.marin.lockout.lockout.texture.TextureProvider;
+import net.minecraft.client.gui.DrawContext;
+import net.minecraft.client.gl.RenderPipelines;
 import net.minecraft.entity.EntityType;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
@@ -18,12 +20,7 @@ public class KillSlimeGoal extends KillMobGoal implements TextureProvider {
         return "Kill Slime";
     }
 
-    @Override
-    public ItemStack getTextureItemStack() {
-        return null;
-    }
-
-    private static final Identifier TEXTURE = Identifier.of(Constants.NAMESPACE, "textures/custom/kill/kill_slime.png");
+    private static final Identifier TEXTURE = Identifier.of(Constants.NAMESPACE, "textures/custom/slime.png");
     @Override
     public Identifier getTextureIdentifier() {
         return TEXTURE;
@@ -32,6 +29,20 @@ public class KillSlimeGoal extends KillMobGoal implements TextureProvider {
     @Override
     public EntityType<?> getEntity() {
         return EntityType.SLIME;
+    }
+    
+    @Override
+    public ItemStack getTextureItemStack() {
+        return null;
+    }
+
+    private static final Identifier OVERLAY = Identifier.of(Constants.NAMESPACE, "textures/custom/kill/kill_x_overlay.png");
+    
+    @Override
+    public boolean renderTexture(DrawContext context, int x, int y, int tick) {
+        context.drawTexture(RenderPipelines.GUI_TEXTURED, TEXTURE, x, y, 0,0, 16, 16, 16, 16);
+        context.drawTexture(RenderPipelines.GUI_TEXTURED, OVERLAY, x, y, 0,0, 16, 16, 16, 16);
+        return true;
     }
 }