+++ src/main/java/me/marin/lockout/lockout/goals/obtain/ObtainDiamondToolsGoal.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public class ObtainDiamondToolsGoal extends ObtainAllItemsGoal {
 
-    private static final List<Item> ITEMS = List.of(Items.DIAMOND_AXE, Items.DIAMOND_HOE, Items.DIAMOND_PICKAXE, Items.DIAMOND_SWORD, Items.DIAMOND_SHOVEL);
+    private static final List<Item> ITEMS = List.of(Items.DIAMOND_AXE, Items.DIAMOND_HOE, Items.DIAMOND_PICKAXE, Items.DIAMOND_SWORD, Items.DIAMOND_SHOVEL, Items.DIAMOND_SPEAR);
 
     public ObtainDiamondToolsGoal(String id, String data) {
         super(id, data);